fix tests, Also of note return on err in schedule CheckExpr
One of the tests was panicing because results was nil, but err was not, so it seems like this error check should have been in place. But I'm not sure why it used to pass
