+++ cmd/bosun/conf/native/invalid/crit-notification-no-template
