+++ cmd/bosun/conf/native/invalid/lookup-key-pairs
