+++ cmd/bosun/conf/native/invalid/number-func-args
