+++ cmd/bosun/conf/native/test.conf
